refactor(delivery): extract EndpointStats + Stats into endpoint_stats.rs (#184)
Frees ~60 lines from endpoint_task.rs (1000 -> 943) so the upcoming
LifecycleAwarePusher wiring fits under the 1000-line CI cap. New
EndpointStats also pre-declares prefetch_fill + last_lifecycle_worst_stage
fields used by the imminent wiring + dashboard surfacing.

No behavior change: pub re-export from endpoint_task keeps existing
import paths working.
