+++ crates/rs-delivery/src/main.rs
@@ -18,6 +18,7 @@ mod disk_cache;
 mod disk_cache_fetcher;
 mod disk_cache_push_sample;
 pub mod endpoint_audit;
+pub mod endpoint_stats;
 pub mod endpoint_task;
 mod ffmpeg_reason;
 mod producer_lag;
